+++ reference/cs/ec-cines/ELCOLOMBIANO.EcCines.Entities.Dtos/Movie/MovieLocationShort.cs
@@ -4,8 +4,8 @@
 /*                Holds the theater part for a movie, where needed.         */
 /* Author:        Carlos Adolfo Ortiz Quirós (COQ)                          */
 /* Date:          Feb.11/2015                                               */
-/* Last Modified: Mar.05/2015                                               */
-/* Version:       1.2                                                       */
+/* Last Modified: Mar.11/2015                                               */
+/* Version:       1.1                                                       */
 /* Copyright (c), 2015 Arkix, El Colombiano                                 */
 /*==========================================================================*/
 
@@ -14,14 +14,12 @@ History
 Feb.11/2015 COQ File created.
 ============================================================================*/
 
-namespace ELCOLOMBIANO.EcCines.Entities.Dtos
-{
+namespace ELCOLOMBIANO.EcCines.Entities.Dtos.Movie {
     /// <summary>
     /// A movie detailed information for JSON serialization. 
     /// Holds the theater part for a movie, where needed. 
     /// </summary>
-    public class MovieLocationShort
-    {
+    public class MovieLocationShort {
         public int id { get; set; }
         public string name { get; set; }        
         public string branchName { get; set; }