+++ reference/cs/ec-cines/ELCOLOMBIANO.EcCines.Entities.Dtos/ParametroSistemaDto.cs
@@ -1,10 +1,8 @@
 ﻿
-namespace ELCOLOMBIANO.EcCines.Entities.Dtos
-{
-    public class ParametroSistemaDto
-    {
-        public int idParametro {get;set;}
-        public string nombreParametro {get;set;}
+namespace ELCOLOMBIANO.EcCines.Entities.Dtos {
+    public class ParametroSistemaDto {
+        public int idParametro { get; set; }
+        public string nombreParametro { get; set; }
         public string valorParametro { get; set; }
         public string descValorParametro { get; set; }
         public char visible { get; set; }